Move Blob to execution datastructures
