+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/execution/versions/eip4844/BlobSchema.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.datastructures.blocks.blob;
+package tech.pegasys.teku.spec.datastructures.execution.versions.eip4844;
 
 import tech.pegasys.teku.infrastructure.ssz.schema.SszPrimitiveSchemas;
 import tech.pegasys.teku.infrastructure.ssz.schema.collections.impl.SszByteVectorSchemaImpl;
